Create Proposal
What should be included in the proposal for presentation #1:
PRODUCT CONTEXT AND NEED
Give an overview of the organization or context for the
system. Identify the purpose. This should serve as a
background to the project/product
PRODUCT FUNCTIONALITY
Summarize the intended major functions the product must
perform or must let the user perform. This should be short
phrases/sentences in list format which provide a high-level
summary. Organize the functions to make them
understandable.
STAKEHOLDERS AND CHARACTERISTICS
1. Identify the key stakeholders of this system.
2. Describe the pertinent characteristics of each
stakeholder.
3. Distinguish the most important stakeholders for
the software product from those who are less
important to satisfy.