Add files via upload
This is a description of the two problems we have identified thus far and our solutions to those problems. Read the document in full please important information is in it, and we need to make a decision on which one we will be trying to implement and begin implementing as soon as possible.